refactor(osn/api): split the auth service monolith into domain modules
services/auth.ts had grown to ~4,500 lines — a single factory closure
returning 53 methods. It is now a services/auth/ directory:

- index.ts composes per-domain module factories (profiles, registration,
  tokens, passkeys, passkey management, profile switch, sessions,
  recovery, security events, step-up, email change, cross-device) over a
  shared AuthContext (config defaults resolved once + injected stores +
  hashIp), and re-exports every previously-public name so all existing
  import paths ("../services/auth") resolve unchanged.
- Cross-module dependencies are explicit factory parameters with acyclic
  layering: context → profiles → tokens → the rest; step-up feeds
  security-events and the step-up-gated flows.
- Stateless pieces move to errors/types/helpers/constants/stores/config
  modules; function bodies are otherwise moved verbatim.
- The three copy-pasted notify*ByAccountId mailers (passkey add/remove,
  cross-device login) collapse into one shared
  notifySecurityEventByAccountId helper with identical metrics + span.

Public surface, wire behaviour, and AuthService type are unchanged;
typecheck and all 706 osn/api tests pass. Wiki path references updated
and the module-directory pattern documented in backend-patterns.
