+++ osn/api/src/services/auth/errors.ts
@@ -1,0 +1,17 @@
+import { Data } from "effect";
+
+// ---------------------------------------------------------------------------
+// Errors
+// ---------------------------------------------------------------------------
+
+export class AuthError extends Data.TaggedError("AuthError")<{
+  readonly message: string;
+}> {}
+
+export class DatabaseError extends Data.TaggedError("DatabaseError")<{
+  readonly cause: unknown;
+}> {}
+
+export class ValidationError extends Data.TaggedError("ValidationError")<{
+  readonly cause: unknown;
+}> {}
