+++ wiki/TODO.md
@@ -600,7 +600,7 @@ Open findings only. Completed fixes archived in [[changelog/security-fixes]].
 - [x] S-M1 (auth) — ~~`pkceStore` unbounded + no expiry sweep~~ — **Obsolete**: `pkceStore` deleted with PKCE cleanup (Phase 5b)
 - [x] S-M2 (auth) — ~~`/authorize` has no rate limiter~~ — **Obsolete**: `/authorize` route deleted with PKCE cleanup (Phase 5b)
 - [ ] S-M4 (auth) — No startup assertion that `OSN_JWT_PRIVATE_KEY` has `sign` usage — assert `key.usages.includes("sign")` after import in `loadJwtKeyPair`
-- [ ] S-L4 (auth) — `logDevOtp` (`osn/api/src/services/auth.ts`) interpolates recipient email + OTP code into a free-text `Effect.logDebug` message, bypassing the annotation-key-based redaction deny-list (`shared/observability/src/logger/redact.ts`). Only fires when `OSN_ENV` is unset/`"local"` (fail-closed env gate verified), so no leak in any deployed env — defence-in-depth note only. If hardening: drop the email from the line (operator already knows the address) — see [[email]], [[observability/overview]]
+- [ ] S-L4 (auth) — `logDevOtp` (`osn/api/src/services/auth/helpers.ts`) interpolates recipient email + OTP code into a free-text `Effect.logDebug` message, bypassing the annotation-key-based redaction deny-list (`shared/observability/src/logger/redact.ts`). Only fires when `OSN_ENV` is unset/`"local"` (fail-closed env gate verified), so no leak in any deployed env — defence-in-depth note only. If hardening: drop the email from the line (operator already knows the address) — see [[email]], [[observability/overview]]
 - [ ] S-L5 (auth) — No boot-time assertion that `OSN_ORIGIN` is set in non-local envs (unlike CORS's `assertCorsOriginsConfigured`). A misconfigured non-local deploy silently falls back to the single default `http://localhost:5173` WebAuthn origin. Fails **closed** (real passkey ceremonies rejected → login breaks), not open, so robustness gap not a security hole. Consider mirroring the CORS boot-assertion for `OSN_ORIGIN` — see [[passkey-primary]]
 - [ ] S-L2 (auth) — Wildcard CORS on `@pulse/api` — restrict to known client origins (mirrors OSN_CORS_ORIGIN pattern) — see [[rate-limiting]]
 - [ ] S-L22 — `listRsvps` counts privacy-filtered rows toward `limit` (weak side-channel oracle)
@@ -665,7 +665,7 @@ Open findings only. Completed fixes archived in [[changelog/performance-fixes]].
 - [ ] P-W2 (zap) — `addMember` fetches all members to check count. Use `COUNT(*)` or catch unique constraint
 - [ ] P-W3 (zap) — `provisionEventChat` non-atomic cross-DB writes
 - [ ] P-W4 (zap) — `getChatMembers` returns all members without pagination
-- [ ] P-I1 (auth) — `logDevOtp` re-reads `process.env.OSN_ENV` on every OTP issuance (registration / step-up / email-change) rather than resolving locality once at startup like `index.ts` does. `process.env` is a backed getter, not a plain read; negligible at these low-QPS rate-limited endpoints. Hoist to a module-level `isLocalEnv` const if touched — `osn/api/src/services/auth.ts`
+- [ ] P-I1 (auth) — `logDevOtp` re-reads `process.env.OSN_ENV` on every OTP issuance (registration / step-up / email-change) rather than resolving locality once at startup like `index.ts` does. `process.env` is a backed getter, not a plain read; negligible at these low-QPS rate-limited endpoints. Hoist to a module-level `isLocalEnv` const if touched — `osn/api/src/services/auth/helpers.ts`
 - [ ] P-I1 (pulse-discovery) — cursor `(start_time, id)` ordering relies on the single-column `events_start_time_idx` for the tiebreak; cheap to add a compound index if series materialisation produces same-second collisions at scale — see `[[event-access]]`
 - [ ] P-I1 (pulse-onboarding) — `getOnboardingStatus` re-validates the JSON `interests` column on every read (defensive walk through `INTEREST_CATEGORIES`) even though `Schema.decodeUnknown(CompleteOnboardingSchema)` already enforced membership at write time. Sub-millisecond cost; flagged purely as Info — only worth changing if profiling later shows it. See [[pulse-onboarding]]
 - [x] P-W2 — `resolvePublicKey` hits DB when `tokenScopes` provided even if `kid` cache is warm. **Fixed** — cache entry now stores `CryptoKey` + `allowedScopes`; scope validated from cache on hit, no DB round-trip — see [[arc-tokens]]
